Update getInitials file to userNames for name helpers, fix CDNAvatar bug
